+++ src/com/vrocketz/spotchu/spot/SpotDao.java
@@ -107,7 +107,7 @@ public class SpotDao {
 
 	  private Spot cursorToSpot(Cursor cursor) {
 	    Spot spot = new Spot();
-	    spot.setId(cursor.getInt(0));
+	    spot.setId(cursor.getLong(0));
 	    spot.setUserId(cursor.getInt(1));
 	    spot.setTag(cursor.getString(2));
 	    spot.setScale(cursor.getInt(3));
